hotel-crm/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
httpx>=0.27.0
redis>=5.0.0
apscheduler>=3.10.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0
